+++ crates/core/component/shielded-pool/src/component.rs
@@ -3,6 +3,7 @@
 mod action_handler;
 mod assets;
 mod fmd;
+mod ics20_withdrawal_with_handler;
 mod metrics;
 mod note_manager;
 mod shielded_pool;
@@ -11,6 +12,7 @@ mod transfer;
 pub use self::metrics::register_metrics;
 pub use assets::{AssetRegistry, AssetRegistryRead};
 pub use fmd::ClueManager;
+pub use ics20_withdrawal_with_handler::Ics20WithdrawalWithHandler;
 pub use note_manager::NoteManager;
 pub use shielded_pool::{ShieldedPool, StateReadExt, StateWriteExt};
 pub use transfer::Ics20Transfer;
